Unneeded scripts deleted
