+++ Assets/Prefabs/Balloon Blue.prefab
@@ -585,10 +585,10 @@ Rigidbody2D:
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
   m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 2
+  m_Mass: 10
+  m_LinearDrag: 5
   m_AngularDrag: 0
-  m_GravityScale: -0.5
+  m_GravityScale: -2
   m_Material: {fileID: 0}
   m_Interpolate: 1
   m_SleepingMode: 1
