+++ Assets/Scripts/Control/ItemController.cs
@@ -1,5 +1,6 @@
 ﻿using Core.Spawner;
 using Core.Spawner.Interfaces;
+using Core.WaiterAsync;
 using Data;
 using Model;
 using UnityEngine;
@@ -9,37 +10,52 @@ using View;
 
 namespace Control
 {
-    public class ItemController : IStartable
+    public class ItemController
     {
         private readonly IPositionGetter _positionGetter;
         private readonly ISpawnerBehaviour _spawnerWithPool;
         private readonly PatternModel _patternModel;
 
-        public ItemController(IPositionGetter positionGetter, ISpawnerBehaviour spawnerWithPool, PatternModel patternModel)
+        private readonly ItemModel _itemModel;
+        private LoopedActionAsync _loopedActionAsync; // Не ижектится на VContainer
+
+        public ItemController(IPositionGetter positionGetter, ISpawnerBehaviour spawnerWithPool, PatternModel patternModel, ItemModel itemModel)
         {
             _positionGetter = positionGetter;
             _spawnerWithPool = spawnerWithPool;
             _patternModel = patternModel;
+            _itemModel = itemModel;
         }
 
-        public void Start()
+        public void StartControl()
         {
             _spawnerWithPool.OnInstantiatedObject += OnSpawned;
+            _spawnerWithPool.Dispose();
+            _loopedActionAsync = new LoopedActionAsync();
+            _loopedActionAsync.DoAction += _spawnerWithPool.Spawn;
+            _loopedActionAsync.Begin(_itemModel.SpawnDuration);
         }
 
         private void OnSpawned(GameObject obj)
         {
             var item = obj.GetComponent<ItemView>();
-            //var pattern = obj.GetComponentInChildren<PatternView>();
+            
+            Assert.IsNotNull(item, "ItemView not found on spawned object in " + obj);
 
             var newPattern = _patternModel.GetPattern(out var type);
-            
-            Assert.IsNotNull(item);
 
+            Assert.IsNotNull(newPattern," Item pattern not created in " + _patternModel);
+
+            // Две методы всегда вызываются одновременно
             item.ChangePosition(_positionGetter.GetRandom());
-            //pattern.ChangePattern(newPattern, type);
-            item.ChangePattern(newPattern, type);
-
+            item.ChangePattern(newPattern, type); 
+        }
+        
+        public void EndControl()
+        {
+            _spawnerWithPool.OnInstantiatedObject -= OnSpawned;
+            _loopedActionAsync.DoAction -= _spawnerWithPool.Spawn;
+            _loopedActionAsync.EndLoop();
         }
     }
 }
