--- Assets/Scripts/Control/SpawnController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4372652f88eeb4c46a2c5713de188480
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
