+++ Assets/Scripts/Core/GameProcess.cs
@@ -9,17 +9,22 @@ namespace Core
 {
     public class GameProcess : IStartable, IDisposable
     {
-        private readonly SpawnController _spawnController;
         private readonly SliderController _sliderController;
         private readonly IScoreControl _scoreControl;
         private readonly EndGameView _endGameView;
+        private readonly ItemController _itemController;
 
-        public GameProcess(SpawnController spawnController, SliderController sliderController, IScoreControl scoreControl, EndGameView endGameView)
+        public GameProcess(
+            SliderController sliderController, 
+            IScoreControl scoreControl, 
+            EndGameView endGameView,
+            ItemController itemController
+            )
         {
-            _spawnController = spawnController;
             _sliderController = sliderController;
             _scoreControl = scoreControl;
             _endGameView = endGameView;
+            _itemController = itemController;
         }
 
         public void Start()
@@ -29,7 +34,7 @@ namespace Core
 
         private void Initialize()
         {
-            _spawnController.StartControl();
+            _itemController.StartControl();
             _sliderController.StartControl();
             _scoreControl.Scored += End;
         }
@@ -41,7 +46,7 @@ namespace Core
 
         private void End()
         {
-            _spawnController.EndControl();
+            _itemController.EndControl();
             _sliderController.EndControl();
             _scoreControl.Scored -= End;
             _endGameView.Activate(RestartGame);
